Actualización de código

+++ src/main/java/com/project/Perseo_Academy/repositories/ICartRepository.java
@@ -3,5 +3,5 @@ package com.project.Perseo_Academy.repositories;
 import com.project.Perseo_Academy.models.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICartRepository extends JpaRepository<Cart, Integer> {
+public interface ICartRepository extends JpaRepository<Cart, Long> {
 }
